requirements.txt
flask>=3.0.0
google-generativeai>=0.7.0
# openai>=1.0.0          # uncomment when needed
# anthropic>=0.25.0      # uncomment when needed
